requirements.txt
google-genai
python-dotenv
pathspec
selenium
webdriver-manager
beautifulsoup4

